% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Committee.getTypes.R
\name{Committee.getTypes}
\alias{Committee.getTypes}
\title{Get the committee types (house, senate, joint, etc.)}
\usage{
Committee.getTypes()
}
\value{
A data frame with a row for each committee type and columns with the following variables describing the committee type:\cr committeeTypes.type*.committeeTypeId,\cr committeeTypes.type*.name.
}
\description{
This function is a wrapper for the Committee.getTypes() method of the PVS API Committee class which returns the existing committee types.
}
\examples{
# First, make sure your personal PVS API key is saved as an option
# (options("pvs.key" = "yourkey")) or in the pvs.key variable:
\dontrun{pvs.key <- "yourkey"}
# get the committee types
\dontrun{comtypes <- Committee.getTypes()}
\dontrun{comtypes}
}
\references{
http://api.votesmart.org/docs/Committee.html\cr
See also: Matter U, Stutzer A (2015) pvsR: An Open Source Interface to Big Data on the American Political Sphere. PLoS ONE 10(7): e0130501. doi: 10.1371/journal.pone.0130501
}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
