% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPVS_key.R
\name{getPVS_key}
\alias{getPVS_key}
\title{Get user's Project Vote Smart API key.}
\usage{
getPVS_key()
}
\description{
This function checks for a "pvs.key" option. If the option exists, its 
stored value is presumed to be the API key. If the option doesn't exist,
the function checks the global environment for a \code{pvs.key} variable 
and uses it as the API key. The function returns the key as a character
string.
}
\examples{
\dontrun{options("pvs.key" = "yourkey")} 
\dontrun{pvs.key <- "yourkey"}
}
\author{
John Bullock <john@johnbullock.org>
}
