% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Office.getTypes.R
\name{Office.getTypes}
\alias{Office.getTypes}
\title{Get all office types tracked}
\usage{
Office.getTypes()
}
\value{
A data frame with rows for each office type and columns with the following variables describing the office type:\cr officeTypes.type*.officeTypeId,\cr officeTypes.type*.officeLevelId,\cr officeTypes.type*.officeBranchId,\cr officeTypes.type*.name.
}
\description{
This function is a wrapper for the Office.getTypes() method of the PVS API Office class which grabs a list of office types Project Vote Smart keeps track of.
}
\examples{
# First, make sure your personal PVS API key is saved as an option
# (options("pvs.key" = "yourkey")) or in the pvs.key variable:
\dontrun{pvs.key <- "yourkey"}
# get all office types tracked
\dontrun{officetypes <- Office.getTypes()}
\dontrun{officetypes}
}
\references{
http://api.votesmart.org/docs/Office.html\cr
See also: Matter U, Stutzer A (2015) pvsR: An Open Source Interface to Big Data on the American Political Sphere. PLoS ONE 10(7): e0130501. doi: 10.1371/journal.pone.0130501
}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
