% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllCities.R
\name{getAllCities}
\alias{getAllCities}
\title{Get basic data on all cities}
\usage{
getAllCities()
}
\value{
A data frame with a row for each city and columns with the following variables describing the city:\cr cities.city*.localId,\cr cities.city*.name,\cr cities.city*.url.
}
\description{
This function is essentially a  wrapper around Local.getCities().
}
\examples{
# First, make sure your personal PVS API key is saved as an option
# (options("pvs.key" = "yourkey")) or in the pvs.key variable:
\dontrun{pvs.key <- "yourkey"}
# get a list of all cities
\dontrun{cities <- getAllCities()}
\dontrun{head(cities)}
}
\references{
http://api.votesmart.org/docs/Local.html\cr
Use State.getStateIDs() to get a list of state IDs.\cr
See also: Matter U, Stutzer A (2015) pvsR: An Open Source Interface to Big Data on the American Political Sphere. PLoS ONE 10(7): e0130501. doi: 10.1371/journal.pone.0130501
}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
