% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllCounties.R
\name{getAllCounties}
\alias{getAllCounties}
\title{Get basic data on all counties}
\usage{
getAllCounties()
}
\value{
A data frame with a row for each county and columns with the following variables describing the county:\cr counties.county*.localId,\cr counties.county*.name,\cr counties.county*.url.
}
\description{
This function is essentially a  wrapper around Local.getCounties().
}
\examples{
# First, make sure your personal PVS API key is saved as character string in the pvs.key variable:
\dontrun{pvs.key <- "yourkey"}
# get a list of all counties 
\dontrun{counties <- getAllCounties()}
\dontrun{head(counties)}
}
\references{
http://api.votesmart.org/docs/Local.html\cr
Use State.getStateIDs() to get a list of state IDs.\cr
See also: Matter U, Stutzer A (2015) pvsR: An Open Source Interface to Big Data on the American Political Sphere. PLoS ONE 10(7): e0130501. doi: 10.1371/journal.pone.0130501
}
\author{
Ulrich Matter <ulrich.matter-at-unibas.ch>
}
